+++ src/wrappers/_ContentErrorBoundary.jsx
@@ -22,6 +22,8 @@ class ContentErrorBoundary extends React.Component {
 
   static getDerivedStateFromError(error) {
     // Update state so the next render will show the fallback UI.
+    // eslint-disable-next-line no-console
+    console.error(error);
     return { hasError: true, error };
   }